DD-2889 Rename vaults
