+++ src/entities/yEarnVaultSavings/loadVaultPool.ts
@@ -1,14 +1,14 @@
 import { Address, log } from "@graphprotocol/graph-ts";
 
-import { Vault } from "../../../generated/schema";
+import { YEarnVault } from "../../../generated/schema";
 
-export function loadVault(poolAddress: Address): Vault {
-  let pool = Vault.load(poolAddress.toHex());
+export function loadVault(poolAddress: Address): YEarnVault {
+  let pool = YEarnVault.load(poolAddress.toHex());
 
   if (!pool) {
     log.error("Vault Pool with address {} not found", [poolAddress.toHex()]);
     throw new Error("Vault Pool not found");
   }
 
-  return pool as Vault;
+  return pool as YEarnVault;
 }
